Documentation
My automation framework that has modules will help the test automation engineer with writing a test automation script, it is a maven project, these framework modules will be declared below. And I apply it on the test case that was sent in the mail. This framework based on POM, Data provider, Object-Oriented Design, and cross-browser concepts.